requirements.txt
requests
python-dotenv
google-generativeai
pyyaml
pandas